+++ src/app/_constant/modal.ts
@@ -1,11 +1,4 @@
 export const MODAL = {
-  POST_FEED: {
-    NAME: "새 게시물 만들기",
-    STEP: {
-      UPLOAD_IMAGE: "이미지 올리기",
-      IMAGE_PREVIEW: "이미지 미리보기",
-      WRITE_CONTENT: "게시글 작성하기",
-    },
-  },
+  POST_FEED: "새 게시물 만들기",
   TEST: "TEST MODAL",
 };
