+++ src/app/_stores/client/modalStore.ts
@@ -5,7 +5,6 @@ const initialState = {
   isOpen: false,
   modalName: "",
   selectedImage: "",
-  modalStep: "",
 };
 
 const useModalStore = create(
@@ -16,7 +15,6 @@ const useModalStore = create(
       closeModal: () => set(() => initialState),
       setModal: (state) => set(() => ({ modalName: state })),
       setSelectedImage: (state) => set(() => ({ selectedImage: state })),
-      setModalStep: (state) => set(() => ({ modalStep: state })),
     }),
     { name: "modal-storage" },
   ),
@@ -26,12 +24,10 @@ interface ModalState {
   isOpen: boolean;
   modalName: string;
   selectedImage: string;
-  modalStep: string;
   openModal: () => void;
   closeModal: () => void;
   setModal: (state: string) => void;
   setSelectedImage: (state: string) => void;
-  setModalStep: (state: string) => void;
 }
 
 export default useModalStore;
